requirements.txt
flask
mysql-connector-python
flask_cors
flask_restful
flask_bcrypt